+++ app/models/part-to-save.ts
@@ -1,0 +1,22 @@
+export interface PartToSave {
+    machine: string;
+    name: string;
+    articleNo: string;
+    tool: string;
+    totalPartsProduced: string;
+    shortShot: string;
+    startupParts: string;
+    burns: string;
+    contaminated: string;
+    oilContaminations: string;
+    smudges: string;
+    deformations: string;
+    damagedInTransport: string;
+    mechanicalDamages: string;
+    scratches: string;
+    flashes: string;
+    silvering: string;
+    removedByRobot: string;
+    airBubbles: string;
+    others: string;
+}
